[UPUP-1084]: add documentation and README
