+++ Example/Example/ContentView.swift
@@ -1,0 +1,19 @@
+//
+//  ContentView.swift
+//  Example
+//
+//  Created by MobileUp on 20.04.2025.
+//
+
+import SwiftUI
+import MUDKit
+
+struct ContentView: View {
+    var body: some View {
+        MUDKitView()
+    }
+}
+
+#Preview {
+    ContentView()
+}
