+++ Example/Example/ExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  ExampleApp.swift
+//  Example
+//
+//  Created by MobileUp on 20.04.2025.
+//
+
+import SwiftUI
+
+@main
+struct ExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
