+++ Package.swift
@@ -7,7 +7,7 @@ let package = Package(
         .iOS(.v15)
     ],
     products: [
-        .library(name: "MUDKit", targets: ["MUDKit"]),
+        .library(name: "MUDKit", targets: ["MUDKit"])
     ],
     dependencies: [
         .package(url: "https://github.com/kean/Pulse", exact: "5.1.3"),
@@ -24,6 +24,6 @@ let package = Package(
                 .product(name: "Alamofire", package: "Alamofire"),
                 .product(name: "KeychainSwift", package: "keychain-swift")
             ]
-        ),
+        )
     ]
 )
