+++ Sources/MUDKit/Configurator/Configurations/EnvironmentConfiguration.swift
@@ -1,10 +1,18 @@
 import Foundation
 
+/// Configuration for managing multiple environments (e.g., dev, prod).
 public struct EnvironmentConfiguration: Sendable {
+    /// List of available environments.
     let environments: [Environment]
+    /// The id of the default environment.
     let defaultEnvironmentId: UUID
+    /// The currently selected environment, if any.
     let selectedEnvironment: Environment?
     
+    /// Initializes the environment configuration with a list of environments and a default one.
+        /// - Parameters:
+        ///   - environments: An array of `Environment` objects representing available environments.
+        ///   - defaultEnvironmentId: The id of the default environment to use if none was selected.
     public init(
         environments: [Environment],
         defaultEnvironmentId: UUID
