+++ Sources/MUDKit/Configurator/MUDKitConfigurator.swift
@@ -1,7 +1,14 @@
-/// Для конфигуратора используется actor, так как компилятор требует  изолировать изменяемое свойство configuration.
+/// A thread-safe configurator for setting up MUDKit with various configurations.
 public actor MUDKitConfigurator {
     public static var configuration: MUDKitConfiguration?
     
+    /// Configures MUDKit with provided settings and returns the resulting configuration.
+        /// - Parameters:
+        ///   - pulseConfiguration: Configuration for Pulse network logging. Optional.
+        ///   - featureToggleConfiguration: Configuration for feature toggles. Optional.
+        ///   - deeplinkConfiguration: Configuration for handling deeplinks. Optional.
+        ///   - environmentConfiguration: Configuration for environment settings. Optional.
+        /// - Returns: The configured `MUDKitConfiguration` or an empty configuration if setup fails.
     public static func setup(
         pulseConfiguration: PulseConfiguration?,
         featureToggleConfiguration: FeatureToggleConfiguration?,
