+++ Sources/MUDKit/Services/MUDKitService.swift
@@ -1,4 +1,8 @@
+/// Utility for interacting with MUDKit features.
 public struct MUDKitService {
+    /// Checks if a feature toggle is enabled.
+        /// - Parameter name: The name of the feature toggle to check.
+        /// - Returns: `true` if the feature toggle is enabled, `false` otherwise.
     public static func isFeatureToggleOn(name: String) -> Bool {
         return MUDKitConfigurator
             .configuration?
