+++ Sources/MUDKit/UI/MUDKitView.swift
@@ -2,6 +2,7 @@ import SwiftUI
 import Pulse
 import PulseUI
 
+/// The main debug interface for MUDKit, providing access to various tools.
 public struct MUDKitView<Content: View>: View {
     private let pulseStore: LoggerStore
     private let pulseMode: ConsoleMode
@@ -45,6 +46,11 @@ public struct MUDKitView<Content: View>: View {
         }
     }
     
+    /// Initializes the MUDKit debug view with optional custom content.
+        /// - Parameters:
+        ///   - pulseStore: The Pulse logger store for network logging. Defaults to `.shared`.
+        ///   - pulseMode: The mode for Pulse console display. Defaults to `.all`.
+        ///   - content: A view builder for custom content to include in the debug interface. Defaults to `EmptyView`.
     public init(
         pulseStore: LoggerStore = .shared,
         pulseMode: ConsoleMode = .all,
